Rename folders/files/vars/reducer path:
omega365->pims
omega365new->omega365

In omeag365 use `typeof ... === 'number'` instead of `Number.isFinite`
