--- src/features/omega365/index.ts
@@ -1,1 +1,0 @@
-export * from "./omega365Slice";
