+++ src/features/pims/index.ts
@@ -1,0 +1,1 @@
+export * from "./pimsSlice";
