+++ src/features/widgets/widgets.tsx
@@ -41,8 +41,8 @@ const Jira = lazy(() => import("features/jira/jira"));
 const Manhole = lazy(() => import("features/manhole/manhole"));
 const XsiteManage = lazy(() => import("features/xsiteManage/xsiteManage"));
 const Offline = lazy(() => import("features/offline/offline"));
+const Pims = lazy(() => import("features/pims/pims"));
 const Omega365 = lazy(() => import("features/omega365/omega365"));
-const Omega365New = lazy(() => import("features/omega365new/omega365"));
 
 export function Widgets() {
     const maximized = useAppSelector(selectMaximized);
@@ -245,10 +245,10 @@ function getWidgetByKey(key: WidgetKey): JSX.Element | string {
             Widget = Offline;
             break;
         case featuresConfig.omegaPims365.key:
-            Widget = Omega365;
+            Widget = Pims;
             break;
-        case featuresConfig.omega365new.key:
-            Widget = Omega365New;
+        case featuresConfig.omega365.key:
+            Widget = Omega365;
             break;
         default:
             return key;
